fix x64 warnings

+++ src/core/Common.cpp
@@ -1850,7 +1850,7 @@ UnicodeString FixedLenDateTimeFormat(const UnicodeString & Format)
   return Result;
 }
 //---------------------------------------------------------------------------
-static UnicodeString FormatTimeZone(long /* Sec */)
+static UnicodeString FormatTimeZone(intptr_t /* Sec */)
 {
   // TTimeSpan Span = TTimeSpan::FromSeconds(Sec);
   UnicodeString Str;
@@ -2129,7 +2129,7 @@ UnicodeString LoadStr(intptr_t Ident, intptr_t MaxLength)
   Result.SetLength(MaxLength > 0 ? MaxLength : 1024);
   HINSTANCE hInstance =  GetGlobalFunctions()->GetInstanceHandle();
   assert(hInstance != 0);
-  intptr_t Length = static_cast<intptr_t>(::LoadString(hInstance, Ident, reinterpret_cast<LPWSTR>(const_cast<wchar_t *>(Result.c_str())), (int)Result.Length()));
+  intptr_t Length = static_cast<intptr_t>(::LoadString(hInstance, (UINT)Ident, reinterpret_cast<LPWSTR>(const_cast<wchar_t *>(Result.c_str())), (int)Result.Length()));
   Result.SetLength(Length);
   return Result;
 }
